Estimate the nonhomgogenous poisson process intensity function from grouped data using a normal kernel

Estimate the nonhomgogenous poisson process intensity function from grouped data using a normal kernel

nd_nhpp_fixed(distances_col, id_col, data = NULL, mu_0 = 0,
  kappa_0 = 1, nu_0 = 1, sigma_0 = 1, alpha = 1, rho = 1,
  L = 4, K = 4, iter_max, warm_up, thin = 1, seed = NULL)

Arguments

distances_col
name of column in data that contains distances

id_col
name of column in data that contains id grouping variable

data
data.frame object that contains grouped distances

mu_0
mean hyperparameter for mu normal base measure. Default is 0; Normal(0,1).

kappa_0
variance hyperparameter for mu normal base measure. Default is 1; Normal(0,1).

nu_0
df hyperparameter for sigma inv chisq base measure. Default is 1; InvChisq(1,1);

sigma_0
scale hyperparameter for sigma inv chisq base measure. Default is 1; InvChisq(1,1);

alpha
outer DP concentration parameter

rho
inner DP concentration parameter

L
component truncation number

K
intensity cluster truncation number

iter_max
total number of iterations for which to run sampler

warm_up
number of iterations for which to burn-in or "warm-up" sampler

thin
number of iterations to thin by

seed
integer with which to initialize random number generator